Add files via upload
Here, I removed the Z score method that I had initially done in the previous repository 'Continued4' and then started the process of optimizing my XGBoost. I put in the hyperparameters to be tuned in the param_grid and then performed Grid Search with Cross Validation (GridSearchCV) in order to find the best parameters for the XGBoost model.